+++ styles/GlobalStyles.ts
@@ -1,0 +1,8 @@
+import { createGlobalStyle } from "styled-components";
+
+const GlobalStyles = createGlobalStyle`
+  body{
+    background-color:${(props) => props.theme.background}
+  }
+`;
+export default GlobalStyles;
